Crate font_atlas_image

Modules

image_atlas_bincode

Enums

DecodingError

EncodingError

Functions

load_atlas
Load the image portion of a font atlas from a slice of bytes, and the metadata portion from a slice.

read_atlas
Load the image portion of a font atlas from one Reader, and the metadata portion from another.

save_atlas
Saves an atlas to two paths. One for the image (using the specified image format), one for the font metadata.

write_atlas
Saves an atlas to two Writers. One for the image (using the specified image format), one for the font metadata.

Type Definitions

DecodingResult

EncodingResult